pyproject.toml
[build-system]
requires = ['setuptools>=64.0.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'PyThinFilm'
version = '0.0.1b0'
description = "Molecular Dynamics protocols for generation of thin film morphologies"
readme = 'README.md'
requires-python = '>=3.6'
license = {text = "GPLv3"}
authors = [{name = "Martin Stroet", email = "m.stroet@uq.edu.au"}]
keywords = []

classifiers = [
    "Environment :: Console",
    "Framework :: Setuptools Plugin",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    'importlib-metadata; python_version<"3.8"',
    'pyyaml>=6.0',
    'click>=8.1',
]

[project.optional-dependencies]
docs = [
    'sphinx>=5.0',
    'sphinx-rtd-theme>=1.0.0',
    'jinja2',
    'numpy',
]

[tool.setuptools.package-data]
'PyThinFilm'= ['templates/**', 'resources/**', "test/*.yml"]


[project.scripts]
pytf = 'PyThinFilm.pytf:parse_command_line'
